Fix bugs and warnings after integrating new thread id system
