+++ include/parlay/internal/block_allocator.h
@@ -20,7 +20,6 @@
 #include <algorithm>
 #include <atomic>
 #include <iostream>
-#include <memory>
 #include <new>
 #include <optional>
