+++ include/parlay/internal/collect_reduce.h
@@ -10,6 +10,7 @@
 #include <utility>
 
 #include "../parallel.h"
+#include "../relocation.h"
 #include "../sequence.h"
 #include "../slice.h"
 #include "../utilities.h"
