+++ include/parlay/internal/delayed/filter.h
@@ -5,6 +5,7 @@
 #include <utility>
 
 #include "../../range.h"
+#include "../../relocation.h"
 #include "../../sequence.h"
 #include "../../utilities.h"
