+++ include/parlay/parallel.h
@@ -74,6 +74,41 @@ inline void blocked_for(size_t start, size_t end, size_t block_size, F&& f, bool
   }
 }
 
+// ----------------------------------------------------------------------------
+//                              Some more extras
+//
+//  - par_do_if :   execute in parallel if do_parallel is true, else sequential
+//  - par_do3 :     parallelize three jobs
+//  - par_do3_if :  guess from the two above
+// ----------------------------------------------------------------------------
+
+template <typename Lf, typename Rf>
+static void par_do_if(bool do_parallel, Lf&& left, Rf&& right, bool cons = false) {
+  if (do_parallel)
+    par_do(std::forward<Lf>(left), std::forward<Rf>(right), cons);
+  else {
+    std::forward<Lf>(left)();
+    std::forward<Rf>(right)();
+  }
+}
+
+template <typename Lf, typename Mf, typename Rf>
+inline void par_do3(Lf&& left, Mf&& mid, Rf&& right) {
+  auto left_mid = [&]() { par_do(std::forward<Lf>(left), std::forward<Mf>(mid)); };
+  par_do(left_mid, std::forward<Rf>(right));
+}
+
+template <typename Lf, typename Mf, typename Rf>
+static void par_do3_if(bool do_parallel, Lf&& left, Mf&& mid, Rf&& right) {
+  if (do_parallel)
+    par_do3(std::forward<Lf>(left), std::forward<Mf>(mid), std::forward<Rf>(right));
+  else {
+    std::forward<Lf>(left)();
+    std::forward<Mf>(mid)();
+    std::forward<Rf>(right)();
+  }
+}
+
 }  // namespace parlay
 
 //****************************************************
